--- a/.travis.yml
+++ b/.travis.yml
@@ -1,31 +1,39 @@
-language: cpp
+language: generic
 
 sudo: false
 
 addons:
   apt:
     sources:
-      - ubuntu-toolchain-r-test
-      - llvm-toolchain-precise-3.5
+     - ubuntu-toolchain-r-test
     packages:
-      - clang-3.5
+     - libstdc++-4.9-dev
 
 matrix:
   include:
      # Linux
      - os: linux
        compiler: clang
-       env: NODE_VERSION="0.12.2" # node abi 14
+       env: NODE_VERSION="0.10" # node abi 11
      - os: linux
        compiler: clang
-       env: NODE_VERSION="0.10.38" # node abi 11
+       env: NODE_VERSION="4" # node abi 46
+     - os: linux
+       compiler: clang
+       env: NODE_VERSION="6" # node abi 48
      # OS X
      - os: osx
        compiler: clang
-       env: NODE_VERSION="0.12.2" # node abi 14
+       env: NODE_VERSION="0.10" # node abi 11
+       osx_image: xcode8.2
      - os: osx
        compiler: clang
-       env: NODE_VERSION="0.10.38" # node abi 11
+       env: NODE_VERSION="4" # node abi 46
+       osx_image: xcode8.2
+     - os: osx
+       compiler: clang
+       env: NODE_VERSION="6" # node abi 48
+       osx_image: xcode8.2
 
 env:
   global:
@@ -33,20 +41,18 @@ env:
     - secure: "mCGuAoX0V75Kl7EBDw5JRZc1KrBglrNdMgFKE7auIl1BFgolHJ0wFodw0F4AZ2uy3SQi3Dfz+yVotpz6NgbLSlqSS1yO66tu0zk0gI3BkaDJ/8A2rfZUNKVEo1nzhM271/9PD0zqS1rUZn5hQPHRwRy7fBNfcXE/CPAvwL8AyBncGB5Iswn2GNett6ffGR+Y7NWBNLePXxiDTBIxVBqBFvvuUschMyuZLdSzKXGMKtKdqVaOihZ3j8tHE7szmnl/wYAO7bT1jhnGnS/K3CcHrZUx+xkppow7UpFD8EcCDgZgs73T1UmVywMaVBvTNznCBlj+qGpsOVbyfo6Xe2DjstBFvJWloTYMoSKwtYNth5rftVnrhKNXdVq+oDuHHtRo/7DnFAfjZErU0QyXBUCBS5MdIoHJrND6U0qODJABBFk86f/4mMEULYzJIYxsp7wP2RpApkGJarvIsaOnAz333ddTeDgamTV6sPka2HbMFgaHQ6lLvc3I4fNdTTMW34xoKVH1D3VKQtAHCeW+L/MUQI6+jFA4Y8XMoh/+IEBXrhh9UayBqjKBYVli/mxmJ7mlq3eMRQMV5hH+kq5KWL3+EzWOXtlyVO7Se5BEqoPuBO5nKKdEtIht3jW9Iy3xnGCVcEEGgg3pOQ41+2yX0vVth8lZ+M/gn04oxTi+1n0heLc="
 
 before_install:
-  - if [[ $(uname -s) == 'Linux' ]]; then
-      export CXX="clang++-3.5";
-      export CC="clang-3.5";
-    fi;
-  - rm -rf ~/.nvm/ && git clone --depth 1 https://github.com/creationix/nvm.git ~/.nvm
-  - source ~/.nvm/nvm.sh
-  - nvm install $NODE_VERSION
-  - nvm use $NODE_VERSION
+ # install clang++ via mason
+ - mkdir /tmp/mason && curl -sSfL https://github.com/mapbox/mason/archive/v0.5.0.tar.gz | tar --gunzip --extract --strip-components=1 --directory=/tmp/mason
+ - /tmp/mason/mason install clang++ 3.9.1
+ - export PATH=$(/tmp/mason/mason prefix clang++ 3.9.1)/bin:${PATH}
+ - export CXX=clang++-3.9
+ - source ~/.nvm/nvm.sh
+ - nvm install $NODE_VERSION
+ - nvm use $NODE_VERSION
 
 install:
-  - npm install --build-from-source --clang=1
+ - npm install --build-from-source --clang=1;
 
 script:
-  - npm test
-
-after_success:
-  - ./test/travis-publish.sh
+ - npm test
+ - ./test/travis-publish.sh