--- a/deps/utf8cpp/utf8/core.h
+++ b/deps/utf8cpp/utf8/core.h
@@ -30,6 +30,12 @@ DEALINGS IN THE SOFTWARE.
 
 #include <iterator>
 
+# ifdef NDEBUG
+#  define FORCE_INLINE inline __attribute__((always_inline))
+# else
+#  define FORCE_INLINE
+# endif
+
 namespace utf8
 {
     // The typedefs for 8-bit, 16-bit and 32-bit unsigned integers
@@ -56,47 +62,47 @@ namespace internal
     const uint32_t CODE_POINT_MAX      = 0x0010ffffu;
 
     template<typename octet_type>
-    inline uint8_t mask8(octet_type oc)
+    FORCE_INLINE uint8_t mask8(octet_type oc)
     {
         return static_cast<uint8_t>(0xff & oc);
     }
     template<typename u16_type>
-    inline uint16_t mask16(u16_type oc)
+    FORCE_INLINE uint16_t mask16(u16_type oc)
     {
         return static_cast<uint16_t>(0xffff & oc);
     }
     template<typename octet_type>
-    inline bool is_trail(octet_type oc)
+    FORCE_INLINE bool is_trail(octet_type oc)
     {
         return ((utf8::internal::mask8(oc) >> 6) == 0x2);
     }
 
     template <typename u16>
-    inline bool is_lead_surrogate(u16 cp)
+    FORCE_INLINE bool is_lead_surrogate(u16 cp)
     {
         return (cp >= LEAD_SURROGATE_MIN && cp <= LEAD_SURROGATE_MAX);
     }
 
     template <typename u16>
-    inline bool is_trail_surrogate(u16 cp)
+    FORCE_INLINE bool is_trail_surrogate(u16 cp)
     {
         return (cp >= TRAIL_SURROGATE_MIN && cp <= TRAIL_SURROGATE_MAX);
     }
 
     template <typename u16>
-    inline bool is_surrogate(u16 cp)
+    FORCE_INLINE bool is_surrogate(u16 cp)
     {
         return (cp >= LEAD_SURROGATE_MIN && cp <= TRAIL_SURROGATE_MAX);
     }
 
     template <typename u32>
-    inline bool is_code_point_valid(u32 cp)
+    FORCE_INLINE bool is_code_point_valid(u32 cp)
     {
         return (cp <= CODE_POINT_MAX && !utf8::internal::is_surrogate(cp));
     }
 
     template <typename octet_iterator>
-    inline typename std::iterator_traits<octet_iterator>::difference_type
+    FORCE_INLINE typename std::iterator_traits<octet_iterator>::difference_type
     sequence_length(octet_iterator lead_it)
     {
         uint8_t lead = utf8::internal::mask8(*lead_it);
@@ -113,7 +119,7 @@ namespace internal
     }
 
     template <typename octet_difference_type>
-    inline bool is_overlong_sequence(uint32_t cp, octet_difference_type length)
+    FORCE_INLINE bool is_overlong_sequence(uint32_t cp, octet_difference_type length)
     {
         if (cp < 0x80) {
             if (length != 1) 
@@ -150,7 +156,7 @@ namespace internal
 
     /// get_sequence_x functions decode utf-8 sequences of the length x
     template <typename octet_iterator>
-    utf_error get_sequence_1(octet_iterator& it, octet_iterator end, uint32_t& code_point)
+    FORCE_INLINE utf_error get_sequence_1(octet_iterator& it, octet_iterator end, uint32_t& code_point)
     {
         if (it == end)
             return NOT_ENOUGH_ROOM;
@@ -161,7 +167,7 @@ namespace internal
     }
 
     template <typename octet_iterator>
-    utf_error get_sequence_2(octet_iterator& it, octet_iterator end, uint32_t& code_point)
+    FORCE_INLINE utf_error get_sequence_2(octet_iterator& it, octet_iterator end, uint32_t& code_point)
     {
         if (it == end) 
             return NOT_ENOUGH_ROOM;
@@ -176,7 +182,7 @@ namespace internal
     }
 
     template <typename octet_iterator>
-    utf_error get_sequence_3(octet_iterator& it, octet_iterator end, uint32_t& code_point)
+    FORCE_INLINE utf_error get_sequence_3(octet_iterator& it, octet_iterator end, uint32_t& code_point)
     {
         if (it == end)
             return NOT_ENOUGH_ROOM;
@@ -195,7 +201,7 @@ namespace internal
     }
 
     template <typename octet_iterator>
-    utf_error get_sequence_4(octet_iterator& it, octet_iterator end, uint32_t& code_point)
+    FORCE_INLINE utf_error get_sequence_4(octet_iterator& it, octet_iterator end, uint32_t& code_point)
     {
         if (it == end)
            return NOT_ENOUGH_ROOM;
@@ -220,7 +226,7 @@ namespace internal
     #undef UTF8_CPP_INCREASE_AND_RETURN_ON_ERROR
 
     template <typename octet_iterator>
-    utf_error validate_next(octet_iterator& it, octet_iterator end, uint32_t& code_point)
+    FORCE_INLINE utf_error validate_next(octet_iterator& it, octet_iterator end, uint32_t& code_point)
     {
         // Save the original value of it so we can go back in case of failure
         // Of course, it does not make much sense with i.e. stream iterators
@@ -272,7 +278,7 @@ namespace internal
     }
 
     template <typename octet_iterator>
-    inline utf_error validate_next(octet_iterator& it, octet_iterator end) {
+    FORCE_INLINE utf_error validate_next(octet_iterator& it, octet_iterator end) {
         uint32_t ignored;
         return utf8::internal::validate_next(it, end, ignored);
     }
@@ -297,13 +303,13 @@ namespace internal
     }
 
     template <typename octet_iterator>
-    inline bool is_valid(octet_iterator start, octet_iterator end)
+    FORCE_INLINE bool is_valid(octet_iterator start, octet_iterator end)
     {
         return (utf8::find_invalid(start, end) == end);
     }
 
     template <typename octet_iterator>
-    inline bool starts_with_bom (octet_iterator it, octet_iterator end)
+    FORCE_INLINE bool starts_with_bom (octet_iterator it, octet_iterator end)
     {
         return (
             ((it != end) && (utf8::internal::mask8(*it++)) == bom[0]) &&
@@ -314,7 +320,7 @@ namespace internal
 	
     //Deprecated in release 2.3 
     template <typename octet_iterator>
-    inline bool is_bom (octet_iterator it)
+    FORCE_INLINE bool is_bom (octet_iterator it)
     {
         return (
             (utf8::internal::mask8(*it++)) == bom[0] &&