--- a/package.json
+++ b/package.json
@@ -15,8 +15,8 @@
   "author": "Andrew Pendleton <andrew@mapbox.com>",
   "license": "MIT",
   "dependencies": {
-    "nan": "2.0.9",
-    "node-pre-gyp": "0.5.x"
+    "nan": "2.4.0",
+    "node-pre-gyp": "0.6.x"
   },
   "bundledDependencies":["node-pre-gyp"],
   "devDependencies": {