--- a/unidecode.js
+++ b/unidecode.js
@@ -3,4 +3,4 @@ var path = require('path');
 var binding_path = binary.find(path.resolve(path.join(__dirname,'./package.json')));
 var binding = require(binding_path);
 
-module.exports = binding.decode;
+module.exports = binding;