This PR optimizes the `decode` function by reducing memory allocations of `std::string` (by calling `string.reserve` and avoiding the creation of an extra string and instead sending the utf8 bytes directly to the `decode` function.

It also adds a new `decodable` function which can be used to test if the string is decodable without incurring the overhead of any string allocation at all.

The `decodable` would speed up this check: https://github.com/mapbox/carmen/blob/4233aab6acde8d7217065ab1f077e07cb90cdadc/lib/util/termops.js#L135

Running `bench.js` before this PR:

```
unidecodeCxx: 688ms
```

And after:

```
cxx decodable: 162ms
cxx decode: 431ms
```

/cc @apendleton @yhahn