@springmeyer 2017-01-27
492c297 improves perf a bit more.

Before:

```
cxx decodable: 163ms
cxx decode: 454ms
```

After:

```
cxx decodable: 130ms
cxx decode: 397ms
```