@springmeyer 2017-01-28
I think this PR is ready. I've got a few other branches going which increase speed more, but they are still experimental. But for the record:

shorty-strings:

```
cxx decode: 330ms
```

shorty-decode:

```
cxx decode2: 275ms
```