@springmeyer 2017-08-01
closing as my understanding from @apendleton is that this module is being deprecated.